requirements.txt
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
lxml==4.6.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
pandas-datareader==0.9.0
Pillow==8.0.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
twilio==6.48.0
urllib3==1.26.2
